+++ docs/reference/changelog.rst
@@ -18,6 +18,11 @@ Changelog
 7.2.0 (unreleased)
 ------------------
 
+Application
+===========
+
+- Add experimental monorepo support for managed builds.
+
 Commands
 ========
 
@@ -28,6 +33,8 @@ Bug fixes
 =========
 
 - Ignore .spread-reuse files when deciding if a source is outdated.
+
+For a complete list of commits, check out the `7.2.0`_ release on GitHub.
 
 7.1.0 (2026-07-07)
 ------------------
@@ -1394,3 +1401,4 @@ For a complete list of commits, check out the `2.7.0`_ release on GitHub.
 .. _7.0.0: https://github.com/canonical/craft-application/releases/tag/7.0.0
 .. _7.0.1: https://github.com/canonical/craft-application/releases/tag/7.0.1
 .. _7.1.0: https://github.com/canonical/craft-application/releases/tag/7.1.0
+.. _7.2.0: https://github.com/canonical/craft-application/releases/tag/7.2.0